# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "User info",
                "operationId": "getInfo",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "405": {
                        "description": "Redirect to Log in"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "User info",
                "operationId": "updateInfo",
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateUserRequest"
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "405": {
                        "description": "Redirect to Log in"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "User info by id",
                "operationId": "getInfoById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "405": {
                        "description": "Redirect to Log in"
                    }
                }
            }
        },
...
        "/api/user/change-password": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "User changepassword",
                "operationId": "changePassword",
                "requestBody": {
                    "$ref": "#/components/requestBodies/ChangePasswordRequest"
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "405": {
                        "description": "Redirect to Log in"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/avatar": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "User set avatar",
                "operationId": "setAvatar",
                "requestBody": {
                    "$ref": "#/components/requestBodies/AvatarRequest"
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "405": {
                        "description": "Redirect to Log in"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/email-verify": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "User verify email letter.",
                "operationId": "verifyEmail",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/task/list": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Tasks list by user",
                "operationId": "getTasks",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/task/{status}/list": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "All user tasks by status",
                "operationId": "getTasksByStatus",
                "parameters": [
                    {
                        "name": "status",
                        "in": "path",
                        "required": true,
                        "example": "waiting"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/task/{id}/offer": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Add new offer to task",
                "operationId": "addOffer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/offer/list": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Offers list by user",
                "operationId": "getOffers",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/offer/remove/{id}": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Remove user offer",
                "operationId": "removeOfferById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/notification/list": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "All notifications by user",
                "operationId": "getNotifications",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/notification/read-all": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Read all user notifications",
                "operationId": "markAllNotificationRead",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/notification/{id}/read": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Read 1 notification by ID",
                "operationId": "markNotificationIdRead",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/chat/{id}/messages": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Chat messages by chat id",
                "operationId": "getChatMessages",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "405": {
                        "description": "Redirect to Log in"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "requestBodies": {
...
            "AvatarRequest": {
                "description": "Set avatar password request.",
                "required": true,
                "content": {
                    "multipart/form-data": {
                        "schema": {
                            "required": [
                                "avatar"
                            ],
                            "properties": {
                                "avatar": {
                                    "description": "New user avatar in base64 format.",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}